+++ fixtures/ios_app_with_helpers/Projects/AppKit/Tests/ios_app_with_helpersKitTests.swift
@@ -1,0 +1,8 @@
+import Foundation
+import XCTest
+
+@testable import ios_app_with_helpersKit
+
+final class ios_app_with_helpersKitTests: XCTestCase {
+
+}