+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 - `ProjectEditor` utility https://github.com/tuist/tuist/pull/702 by @pepibumur.
 - Fix warnings in the project, refactor SHA256 diegest code https://github.com/tuist/tuist/pull/704 by @rowwingman.
 - Define `ArchiveAction` on `Scheme` https://github.com/tuist/tuist/pull/697 by @grsouza.
+- `tuist edit` command https://github.com/tuist/tuist/pull/703 by @pepibumur.
 
 ## 0.19.0